requirements.txt
attr
ConfigParser
cryptography
docutils
filelock
HTMLParser
ipython
ipywidgets
keyring
pandas
protobuf
pyOpenSSL
python-dotenv
redis
Sphinx
thread
typing_extensions
urllib3_secure_extra
google-cloud-storage
requests
flask
pytz